Cleaned api for schema, deleted exploratory custom json editor, added packages for more out of box solution
