+++ viewer/package.json
@@ -44,6 +44,7 @@
     "react-cool-dimensions": "^3.0.0",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
+    "react-jsonschema-form": "^1.8.1",
     "react-markdown": "^9.0.1",
     "react-plotly.js": "^2.6.0",
     "react-router-dom": "^6.26.2",
